Move map marker method out of app controller
